prepping for azure repo

+++ ScrumPoker-react/Models/GameModel.cs
@@ -6,7 +6,7 @@ public class GameModel
     public IList<int> VotingSystem { get; set; }
     public IList<int> VotingCardsTopRow { get; set; }
     public IList<int> VotingCardsBottomRow { get; set; }
-    public int AverageScore { get; set; }
+    public double AverageScore { get; set; }
     public bool ScoresDisplayed { get; set; }
